fix(connect): isolate diagnostic contract from Den runtime
