+++ ee/apps/den-api/src/connect-diagnostics.ts
@@ -1,14 +1,22 @@
-import { createHmac, randomUUID } from "node:crypto"
 import {
-  CONNECT_DIAGNOSTIC_CLIENT_HEADER,
   connectDiagnosticIncidentBatchSchema,
-  type ConnectDiagnosticClientEvent,
   type ConnectDiagnosticIncident,
   type ConnectDiagnosticOutcome,
-  type ConnectDiagnosticPhase,
 } from "@openwork/types/den/connect-diagnostics"
 import { env } from "./env.js"
 import { appLogger } from "./observability/logger.js"
+import {
+  connectDiagnosticClientId,
+  denMcpDiagnosticIncident as createDenMcpDiagnosticIncident,
+  desktopConnectDiagnosticIncidents,
+  pseudonymizeConnectDiagnosticValue,
+} from "./connect-diagnostic-contract.js"
+
+export {
+  connectDiagnosticClientId,
+  desktopConnectDiagnosticIncidents,
+  pseudonymizeConnectDiagnosticValue,
+} from "./connect-diagnostic-contract.js"
 
 const CONNECT_DIAGNOSTIC_DELIVERY_TIMEOUT_MS = 3_000
 
@@ -19,51 +27,10 @@ export class ConnectDiagnosticDeliveryError extends Error {
   }
 }
 
-export function pseudonymizeConnectDiagnosticValue(
-  kind: "organization" | "client",
-  value: string,
-  secret: string,
-): string {
-  return createHmac("sha256", secret)
-    .update("openwork-connect-diagnostics-v1\0")
-    .update(kind)
-    .update("\0")
-    .update(value.trim())
-    .digest("hex")
-}
-
 function diagnosticsConfiguration(): { bearerToken: string; origin: string } | null {
   const bearerToken = env.diagnostics.bearerToken?.trim() ?? ""
   const origin = env.diagnostics.origin?.trim().replace(/\/+$/u, "") ?? ""
   return bearerToken && origin ? { bearerToken, origin } : null
-}
-
-export function desktopConnectDiagnosticIncidents(input: {
-  organizationId: string
-  events: readonly ConnectDiagnosticClientEvent[]
-  bearerToken: string
-}): ConnectDiagnosticIncident[] {
-  const organizationHash = pseudonymizeConnectDiagnosticValue(
-    "organization",
-    input.organizationId,
-    input.bearerToken,
-  )
-  return input.events.map((event) => {
-    const { clientId, ...metadata } = event
-    return {
-      ...metadata,
-      source: "desktop" as const,
-      organizationHash,
-      clientHash: pseudonymizeConnectDiagnosticValue("client", clientId, input.bearerToken),
-    }
-  })
-}
-
-function phaseForMcpMethod(method: string | null): ConnectDiagnosticPhase {
-  if (method === "initialize") return "initialize"
-  if (method === "notifications/initialized") return "initialized_notice"
-  if (method === "tools/list") return "tools_list"
-  return "mcp_request"
 }
 
 export function denMcpDiagnosticIncident(input: {
@@ -78,36 +45,10 @@ export function denMcpDiagnosticIncident(input: {
   errorCode: string | null
   bearerToken: string
 }): ConnectDiagnosticIncident {
-  return {
-    schemaVersion: 1,
-    eventId: randomUUID(),
-    attemptId: null,
-    source: "den",
-    observedAt: input.observedAt,
-    organizationHash: pseudonymizeConnectDiagnosticValue(
-      "organization",
-      input.organizationId,
-      input.bearerToken,
-    ),
-    clientHash: input.clientId
-      ? pseudonymizeConnectDiagnosticValue("client", input.clientId, input.bearerToken)
-      : null,
-    phase: phaseForMcpMethod(input.method),
-    outcome: input.outcome,
-    errorCode: input.errorCode,
-    networkCode: null,
-    httpStatus: input.httpStatus,
-    retryable: input.httpStatus === null ? null : input.httpStatus >= 500,
-    deviceOnline: null,
-    durationMs: Math.max(0, Math.round(input.durationMs)),
-    consecutiveFailures: input.outcome === "failure" ? 1 : 0,
-    maintenanceAttempt: null,
-    appVersion: null,
-    platform: null,
+  return createDenMcpDiagnosticIncident({
+    ...input,
     serverVersion: env.serviceVersion?.trim() || null,
-    engineVersion: null,
-    serverRequestId: input.requestId,
-  }
+  })
 }
 
 export async function forwardConnectDiagnosticIncidents(input: {
@@ -157,13 +98,6 @@ export function reportDenMcpDiagnostic(input: Omit<Parameters<typeof denMcpDiagn
       error_type: error instanceof Error ? error.name : typeof error,
     })
   })
-}
-
-export function connectDiagnosticClientId(request: Request): string | null {
-  const value = request.headers.get(CONNECT_DIAGNOSTIC_CLIENT_HEADER)?.trim() ?? ""
-  return /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/iu.test(value)
-    ? value
-    : null
 }
 
 export function configuredConnectDiagnosticBearerToken(): string | null {
